requirements.txt
opencv-python
scikit-learn
numpy
seaborn
matplotlib
tensorflow
albumentations
tqdm
natsor
pillow
keras
flask
werkzeug
